==> 1_mergePlayerIds.ipynb <==
# Mapping Player IDs
The aim of this notebook is to create a one to one mapping between the player IDs from Kaggle Data and CMU data.
This is done by matching the first name, last name and the birth year of each player.

import pandas as pd
import numpy as np

# Two player sources in Kaggle Data
kagglePlayerData1 = pd.read_csv("Kaggle Data/player.csv")
kagglePlayerData2 = pd.read_csv("Kaggle Data/common_player_info.csv")
# Players from CMU Data
cmuPlayerData = pd.read_csv("Data/players.txt")

# convert kaggle player data's firstname lastname to lowercase
kagglePlayerData1['first_name'] = kagglePlayerData1['first_name'].str.strip().str.lower()
kagglePlayerData1['last_name'] = kagglePlayerData1['last_name'].str.strip().str.lower()
kagglePlayerData1["id"].nunique()

kagglePlayerData2["birth_year"] = kagglePlayerData2.apply(lambda x: x["birthdate"][0:4], axis=1)
kagglePlayerData2['first_name'] = kagglePlayerData2['first_name'].str.strip().str.lower()
kagglePlayerData2['last_name'] = kagglePlayerData2['last_name'].str.strip().str.lower()
kagglePlayerData2['person_id'].nunique()

# convert CMU player data's firstname lastname to lowercase
cmuPlayerData['firstname'] = cmuPlayerData['firstname'].str.lower().str.strip()
cmuPlayerData['lastname'] = cmuPlayerData['lastname'].str.lower().str.strip()
cmuPlayerData = cmuPlayerData.rename(columns={'firstname': 'first_name'})
cmuPlayerData = cmuPlayerData.rename(columns={'lastname': 'last_name'})
cmuPlayerData["birth_year"] = cmuPlayerData.apply(lambda x: str(x["birthdate"])[0:4], axis=1)
cmuPlayerData['ilkid'].nunique()

# merge both the dataframes
# mergedPlayers = pd.merge(cmuPlayerData, kagglePlayerData1, left_on=['firstname', 'lastname'], right_on=['first_name','last_name'])
'''
    Using the 1st datasource from Kaggle we were not able to create a one to one mapping since it
    did not contain birth year and there were many players having the same first and last name.
'''

mergedPlayers = cmuPlayerData.merge(kagglePlayerData2, on=['first_name', 'last_name','birth_year'])
print("Total rows = "+str(len(mergedPlayers.index)))
print("Unique ilkid = "+ str(mergedPlayers["ilkid"].nunique()))
print("Unique id = "+ str(mergedPlayers["person_id"].nunique()))

mergedPlayers

Out of 3572 players from the CMU data, we were only able to map 2717 players with the Kaggle data. We speculate that either the names or the birth year had some discrepancy or the player just did not exit in the Kaggle data.

# Find out which players did not merge
merged_df = pd.merge(cmuPlayerData, mergedPlayers, on='ilkid', how='left', indicator=True)
rows_only_in_playerOfDf = merged_df[merged_df['_merge'] == 'left_only']
rows_only_in_playerOfDf


mergedPlayers.to_csv("merged_players.csv")
