==> 3_gameDataPreProcessing.ipynb <==
# Game Data preprocessing
Create a dataframe that has team and its players data in a game.

import pandas as pd
import numpy as np

gameToPlayerDf = pd.read_csv("gameToPlayer.csv")
gameDf = pd.read_csv("Kaggle Data/game.csv")

# Group players based on team id and game id so that we get individual teams and their players
grouped_df = gameToPlayerDf.groupby(['game_id','team_id'])['player_id'].apply(lambda x: list(set(x))).reset_index()
grouped_df = grouped_df.sort_values(by="game_id")
grouped_df.sort_values(by="game_id")

# Adding winning team name in the game DF
gameDf['winning_team_abbr'] = gameDf.apply(lambda row: row['team_abbreviation_home'] if row['wl_home'] == 'W' else row['team_abbreviation_away'], axis=1)
teamWinDf = gameDf[["game_id","winning_team_abbr"]]
teamWinDf.sort_values(by="game_id")

# Merging with teams data from Kaggle to get the nicknames of the team.
teams = pd.read_csv("Kaggle data/team.csv")
teamWinDf = pd.merge(teamWinDf, teams, left_on='winning_team_abbr', right_on='abbreviation', how='left')[['game_id', 'winning_team_abbr', 'nickname']]
teamWinDf

# Append winning team name in the grouped_df object
teamsWithWinDf = teamWinDf.merge(grouped_df, on="game_id")
teamsWithWinDf

teamsWithWinDf.to_csv("teamsWithWin.csv")
