==> 4_mappingPlayersToGames.ipynb <==
# Mapping games to players
Mapping game data from gameDataPreProcessing.ipynb to player data from mergedPlayer.ipynb

import pandas as pd
import numpy as np

playersDf = pd.read_csv("merged_players.csv")[['ilkid','person_id','playercode','team_id']]
teamsWithWinDf = pd.read_csv("teamsWithWin.csv")

playersDf["person_id"] = playersDf["person_id"].astype(str)

# Now explode the dataframe
teamsWithWinDf.player_id = teamsWithWinDf.player_id.str[1:-1].str.split(',').tolist()
expandedTeamsWithWinDf = teamsWithWinDf.explode('player_id').reset_index()
expandedTeamsWithWinDf['player_id'] = expandedTeamsWithWinDf['player_id'].apply(lambda x: x.strip())
expandedTeamsWithWinDf['game_id'].nunique() 


# Now join with the player stats
joinedTable = pd.merge(expandedTeamsWithWinDf, playersDf, left_on="player_id", right_on="person_id")
joinedTable['game_id'].nunique() 

games = pd.read_csv("Kaggle data/game.csv")[['game_id','game_date']]
joinedTable = pd.merge(joinedTable, games, on='game_id', how='left')
joinedTable

joinedTable = joinedTable[['game_id', 'game_date', 'winning_team_abbr', 'team_id_x', 'team_id_y', 'player_id', 'ilkid', 'playercode']]
joinedTable.to_csv("player_with_game_id.csv")
