==> 5_normalizingPlayerStat.ipynb <==
# Creating normalized and cumilative player statistics
The player statistics in the CMU data are recorded for the entire season that is the points scored by a player in all the games that they played in that season, we are interested in knowing the statistics per game for a player. If we use the raw data for a season, players which have player more games will have an edge over the players which have played less games. Therefore we normalize these statistics by dividing these attributes by the number of games player each season. Once this is done, for every year we want an average of these statistics including the past years.

import pandas as pd

gamesDf = pd.read_csv("Kaggle Data/game.csv")

playerRegularSeason = pd.read_csv("Data/player_regular_season.txt")
# playerPlayoffs = pd.read_csv("Data/player_playoffs.txt")
# playerPlayoffs["ilkid"] = playerPlayoffs["ilkid"].str.strip()

# new_rows = []
# for i, row in playerPlayoffs.iterrows():
#     player = playerRegularSeason[(playerRegularSeason["ilkid"] == row["ilkid"]) & (playerRegularSeason["year"] == row["year"]) & (playerRegularSeason["team"] == row["team"])]
#     if len(player.index) == 0:
#         new_rows.append(row.to_dict())

# print(new_rows[0]["ilkid"])
# print(playerRegularSeason.at[0,'ilkid'])
# print(len(new_rows))

# new_df = pd.DataFrame(new_rows)
# playerRegularSeason = pd.concat([playerRegularSeason, new_df],ignore_index=True)
# # playerRegularSeason = playerRegularSeason.append(new_df, ignore_index=True)
# playerRegularSeason

# Normalize the player stats by dividing stats by games played
playerRegularSeason1 = playerRegularSeason.iloc[:,:7]
playerRegularSeason2 = playerRegularSeason.iloc[:,7:]
playerRegularSeason2Nomarlized = playerRegularSeason2.div(playerRegularSeason1["gp"], axis=0).round(2)
playerRegularSeason2Nomarlized

# Rejoin and form the full table
playerRegularSeasonNormalized = pd.concat([playerRegularSeason1, playerRegularSeason2Nomarlized], axis=1)
playerRegularSeasonNormalized

playerRegularSeasonNormalized.to_csv("player_regular_season_normalized.csv")

playerRegularSeasonNormalized = playerRegularSeasonNormalized[playerRegularSeasonNormalized["year"] >= 1995]
playerRegularSeasonNormalized

# I want to calculate a cummalative table where a player stats for 1996 for ex is the stat of 1995
playerRegularSeasonNormalized.sort_values("year", inplace=True)

# Get all columns for cummalative sume
column_names = playerRegularSeasonNormalized.columns.to_list()
relevant_columns = column_names[6:]

# Find cummalative sum for each player for each stat
playerRegularSeasonNormalized['count'] = playerRegularSeasonNormalized.groupby('ilkid').cumcount() + 1
for col in relevant_columns:
    playerRegularSeasonNormalized[f'{col}_cumsum'] = playerRegularSeasonNormalized.groupby('ilkid')[col].cumsum()
    playerRegularSeasonNormalized[f'avg_{col}'] = playerRegularSeasonNormalized[f'{col}_cumsum'] / playerRegularSeasonNormalized['count']
playerRegularSeasonNormalized

playerRegularSeasonNormalized.to_csv("cumilative_player_attributes.csv")
