==> 9_L2-norm.ipynb <==
import pandas as pd
from sklearn.preprocessing import StandardScaler
import numpy as np
games = pd.read_csv("prev_year_team_attr_by_game.csv", index_col=0)

def getL2Norm(arr):
  return np.linalg.norm(arr)

y_test = games["winner"]

count = 0
y_pred = []
for i, row in games.iterrows():
  t1_score = getL2Norm(row[2:19])
  t2_score = getL2Norm(row[19:36])
  winner = row[-1]
  if t1_score > t2_score:
    y_pred.append(0)
  else:
    y_pred.append(1)

# print("Accuracy = ",count/len(games.index))

len(y_pred)
len(y_test)

from sklearn.metrics import confusion_matrix
from sklearn.metrics import classification_report

conf_matrix = confusion_matrix(y_test, y_pred)
# print("Test Accuracy = ", accuracy)
print(classification_report(y_test,y_pred))

conf_matrix
