==> 10_Sequention_NN.ipynb <==
import pandas as pd
from sklearn.preprocessing import StandardScaler

games = pd.read_csv("prev_year_team_attr_by_game.csv", index_col=0)


games["t1_fg"] = games["t1_fgm"]/games["t1_fga"]
games["t2_fg"] = games["t2_fgm"]/games["t2_fga"]

games["t1_ft"] = games["t1_ftm"]/games["t1_fta"]
games["t2_ft"] = games["t2_ftm"]/games["t2_fta"]

games["t1_tp"] = games["t1_tpm"]/games["t1_tpa"]
games["t2_tp"] = games["t2_tpm"]/games["t2_tpa"]

games.drop('t1_fgm', axis=1, inplace=True)
games.drop('t1_fga', axis=1, inplace=True)

games.drop('t2_fgm', axis=1, inplace=True)
games.drop('t2_fga', axis=1, inplace=True)


games.drop('t1_ftm', axis=1, inplace=True)
games.drop('t1_fta', axis=1, inplace=True)

games.drop('t2_ftm', axis=1, inplace=True)
games.drop('t2_fta', axis=1, inplace=True)


games.drop('t1_tpm', axis=1, inplace=True)
games.drop('t1_tpa', axis=1, inplace=True)

games.drop('t2_tpm', axis=1, inplace=True)
games.drop('t2_tpa', axis=1, inplace=True)

games.drop('t1_reb', axis=1, inplace=True)
games.drop('t2_reb', axis=1, inplace=True)


y = games["winner"]


games.drop('winner', axis=1, inplace=True)
games.drop('t1_team_ilkid', axis=1, inplace=True)
games.drop('t2_team_ilkid', axis=1, inplace=True)

X = games[games.columns[2:]]

scaler = StandardScaler()
scaled_data = scaler.fit_transform(X)
X = pd.DataFrame(scaled_data, columns=X.columns)

X.fillna(0, inplace=True)

from sklearn.neural_network import MLPClassifier
from sklearn.model_selection import train_test_split
from sklearn.metrics import accuracy_score
from sklearn.metrics import confusion_matrix
from sklearn.metrics import classification_report

X_train, X_test, y_train, y_test = train_test_split(X, y, test_size=0.2)

from keras.models import Sequential
from keras.layers import Dense

model = Sequential([
    Dense(68, activation='relu', input_shape=(X_train.shape[1],)),
    Dense(34, activation='relu'),
    Dense(1, activation='sigmoid')
])

model.compile(optimizer='adam', loss='binary_crossentropy', metrics=['accuracy'])

# Train the model
history = model.fit(X_train, y_train, epochs=15, batch_size=32, validation_split=0.2)

# Evaluate the model
train_loss, train_accuracy = model.evaluate(X_train, y_train)
print(f'Loss: {train_loss}, Accuracy: {train_accuracy}')

y_pred = model.predict(X_test)
y_pred_binary = (y_pred > 0.5).astype(int)
accuracy = accuracy_score(y_test, y_pred_binary)
conf_matrix = confusion_matrix(y_test, y_pred_binary)
print("Test Accuracy = ", accuracy)
print(classification_report(y_test,y_pred_binary))

conf_matrix
